Add config test button to running backup
