+++ app/views/fluentd/settings/running_backup/show.html.haml
@@ -1,6 +1,11 @@
-- page_title t('.page_title', label: @fluentd.label) do
-  - if @backup_file.content
-    - link_to reuse_daemon_setting_running_backup_path, method: 'post', class: "btn btn-primary pull-right" do
+- page_title t('.page_title', label: @fluentd.label)
+
+- if @backup_file.content
+  %p{class: "pull-right"}
+    = link_to configtest_daemon_setting_running_backup_path, method: 'post', class: "btn btn-default" do
+      = icon('fa-legal')
+      = t("terms.configtest")
+    = link_to reuse_daemon_setting_running_backup_path, method: 'post', class: "btn btn-primary" do
       = icon('fa-pencil')
       = t("terms.reuse")
