+++ spec/features/fluentd/setting/ranning_backup_spec.rb
@@ -43,6 +43,42 @@ describe "running_backup", stub: :daemon do
         page.should have_text(I18n.t('messages.config_successfully_copied', brand: 'fluentd') )
         page.should have_text(backup_content)
       end
+
+      describe "configtest" do
+        let(:backup_content){ config }
+        let(:daemon) { build(:fluentd, variant: "fluentd_gem") } # To use fluentd_gem for real dry-run checking
+        before do
+          click_link I18n.t("terms.configtest")
+        end
+
+        context "invalid configfile" do
+          let(:config) { <<-CONFIG }
+          <source>
+            type aaaaaaaaaaaa
+          </source>
+          CONFIG
+
+          it do
+            page.should_not have_css('.alert-success')
+            page.should have_css('.alert-danger')
+            page.should have_text(%Q|Unknown input plugin 'aaaaaaaaaaaa'|)
+          end
+        end
+
+        context "valid configfile" do
+          let(:config) { <<-CONFIG }
+          <source>
+            type syslog
+            tag syslog
+          </source>
+          CONFIG
+
+          it do
+            page.should have_css('.alert-success')
+            page.should_not have_css('.alert-danger')
+          end
+        end
+      end
     end
   end
 end
